dom/indexedDB/test/test_file_transaction_abort.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>Indexed Database Property Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    const READ_WRITE = "readwrite";

    const name = window.location.pathname;

    const objectStoreName = "Blobs";

    const fileData = { key: 1, file: getRandomFile("random.bin", 100000) };

    {
      let request = indexedDB.open(name, 1);
      request.onerror = errorHandler;
      request.onupgradeneeded = grabEventAndContinueHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      is(event.type, "upgradeneeded", "Got correct event type");

      let db = event.target.result;
      db.onerror = errorHandler;

      let objectStore = db.createObjectStore(objectStoreName, { });

      event = yield undefined;

      is(event.type, "success", "Got correct event type");

      let trans = db.transaction([objectStoreName], READ_WRITE);
      objectStore = trans.objectStore(objectStoreName);

      request = objectStore.add(fileData.file, fileData.key);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.target.result, fileData.key, "Got correct key");

      trans.onabort = grabEventAndContinueHandler;
      trans.abort();
      event = yield undefined;

      is(event.type, "abort", "Got correct event type");

      is(getFileDBRefCount(name, 1), 0, "Correct db ref count");

      fileData.file = null;
    }

    scheduleGC();
    yield undefined;

    ok(!hasFileInfo(name, 1), "Correct ref count");

    finishTest();
    yield undefined;
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
